+++ pepper-sync/src/sync.rs
@@ -3,7 +3,7 @@
 use std::collections::{BTreeMap, HashMap};
 use std::ops::Range;
 use std::sync::Arc;
-use std::sync::atomic::{self, AtomicBool, AtomicU8};
+use std::sync::atomic::{self, AtomicBool, AtomicU8, AtomicU32};
 use std::time::{Duration, SystemTime};
 
 use tokio::sync::{RwLock, mpsc};
@@ -345,7 +345,7 @@ where
     let (mempool_transaction_sender, mut mempool_transaction_receiver) = mpsc::channel(100);
     let shutdown_mempool = Arc::new(AtomicBool::new(false));
     let shutdown_mempool_clone = shutdown_mempool.clone();
-    let unprocessed_mempool_transactions_count = Arc::new(AtomicU8::new(0));
+    let unprocessed_mempool_transactions_count = Arc::new(AtomicU32::new(0));
     let unprocessed_mempool_transactions_count_clone =
         unprocessed_mempool_transactions_count.clone();
     let mempool_handle = tokio::spawn(async move {
@@ -359,23 +359,26 @@ where
     });
 
     // pre-scan initialisation
-    let mut wallet_guard = wallet.write().await;
-
     let chain_height = client::get_chain_height(fetch_request_sender.clone()).await?;
     if chain_height == 0.into() {
         return Err(SyncError::ServerError(ServerError::GenesisBlockOnly));
     }
-    let last_known_chain_height =
-        checked_wallet_height(&mut *wallet_guard, chain_height, consensus_parameters)?;
+    let last_known_chain_height = checked_wallet_height(
+        &mut *wallet.write().await,
+        chain_height,
+        consensus_parameters,
+    )?;
 
-    let ufvks = wallet_guard
+    let ufvks = wallet
+        .read()
+        .await
         .get_unified_full_viewing_keys()
         .map_err(SyncError::WalletError)?;
 
     #[cfg(not(feature = "darkside_test"))]
     transparent::update_addresses_and_scan_targets(
         consensus_parameters,
-        &mut *wallet_guard,
+        wallet.clone(),
         fetch_request_sender.clone(),
         &ufvks,
         last_known_chain_height,
@@ -388,14 +391,14 @@ where
     update_subtree_roots(
         consensus_parameters,
         fetch_request_sender.clone(),
-        &mut *wallet_guard,
+        &mut *wallet.write().await,
     )
     .await?;
 
     add_initial_frontier(
         consensus_parameters,
         fetch_request_sender.clone(),
-        &mut *wallet_guard,
+        &mut *wallet.write().await,
     )
     .await?;
 
@@ -404,21 +407,19 @@ where
         fetch_request_sender.clone(),
         last_known_chain_height,
         chain_height,
-        &mut *wallet_guard,
+        &mut *wallet.write().await,
     )
     .await?;
 
     state::set_initial_state(
         consensus_parameters,
         fetch_request_sender.clone(),
-        &mut *wallet_guard,
+        &mut *wallet.write().await,
         chain_height,
     )
     .await?;
 
-    expire_transactions(&mut *wallet_guard)?;
-
-    drop(wallet_guard);
+    expire_transactions(&mut *wallet.write().await)?;
 
     // create channel for receiving scan results and launch scanner
     let (scan_results_sender, mut scan_results_receiver) = mpsc::unbounded_channel();
@@ -730,8 +731,14 @@ where
         / (total_blocks - sync_state.initial_sync_state.previously_scanned_blocks) as f32)
         * 100.0)
         .clamp(0.0, 100.0);
+    if percentage_session_blocks_scanned.is_nan() {
+        percentage_session_blocks_scanned = 100.0;
+    }
     let mut percentage_total_blocks_scanned =
         ((total_blocks_scanned as f32 / total_blocks as f32) * 100.0).clamp(0.0, 100.0);
+    if percentage_total_blocks_scanned.is_nan() {
+        percentage_total_blocks_scanned = 100.0;
+    }
 
     let session_sapling_outputs_scanned = total_sapling_outputs_scanned
         - sync_state
@@ -752,8 +759,14 @@ where
         / (total_outputs - previously_scanned_outputs) as f32)
         * 100.0)
         .clamp(0.0, 100.0);
+    if percentage_session_outputs_scanned.is_nan() {
+        percentage_session_outputs_scanned = 100.0;
+    }
     let mut percentage_total_outputs_scanned =
         ((total_outputs_scanned as f32 / total_outputs as f32) * 100.0).clamp(0.0, 100.0);
+    if percentage_total_outputs_scanned.is_nan() {
+        percentage_total_outputs_scanned = 100.0;
+    }
 
     if sync_state.scan_ranges().iter().any(|scan_range| {
         scan_range.priority() == ScanPriority::ScannedWithoutMapping
@@ -965,7 +978,7 @@ pub fn set_transactions_failed(
 /// Returns true if the scanner and mempool are shutdown.
 fn is_shutdown<P>(
     scanner: &Scanner<P>,
-    mempool_unprocessed_transactions_count: Arc<AtomicU8>,
+    mempool_unprocessed_transactions_count: Arc<AtomicU32>,
 ) -> bool
 where
     P: consensus::Parameters + Sync + Send + 'static,
@@ -1818,7 +1831,7 @@ where
 async fn mempool_monitor<C>(
     mut client: C,
     mempool_transaction_sender: mpsc::Sender<RawTransaction>,
-    unprocessed_transactions_count: Arc<AtomicU8>,
+    unprocessed_transactions_count: Arc<AtomicU32>,
     shutdown_mempool: Arc<AtomicBool>,
 ) -> Result<(), MempoolError>
 where
@@ -1838,10 +1851,18 @@ where
                         mempool_stream_message = mempool_stream.message() => {
                             match mempool_stream_message.unwrap_or(None) {
                                 Some(raw_transaction) => {
-                                     let _ignore_error = mempool_transaction_sender
+                                     match mempool_transaction_sender
                                         .send(raw_transaction)
-                                        .await;
-                                    unprocessed_transactions_count.fetch_add(1, atomic::Ordering::Release);
+                                        .await {
+                                            Ok(_) => {
+                                                unprocessed_transactions_count.fetch_add(1, atomic::Ordering::Release);
+                                            }
+                                            Err(_) => {
+                                                unprocessed_transactions_count.store(0, atomic::Ordering::Release);
+                                                shutdown_mempool.store(true, atomic::Ordering::Release);
+                                                break 'main;
+                                            }
+                                        }
                                 }
                                 None => {
                                     continue 'main;